requirements.txt
django==5.0.6
lib-version-remla24-team02
python-decouple==3.8
requests==2.31.0
setuptools==69.5.1
prometheus-client==0.20.0